tablet upload working

+++ push_settings.sh
@@ -1,15 +1,20 @@
 #!/bin/bash
-set -xe
-rooms="study kitchen lounge hallwaylarge bathroommain bathroomguest bedroom1 bedroom2 bedroom3"
+set -e
+
+rooms="study lounge hallwaylarge bathroomguest bedroom1 bedroom2 bedroom3 kitchen bathroommain"
 
 for room in $rooms; do
-  echo "Deploying to $room"
-  adb connect tablet-$room.imagilan
-  adb -s tablet-$room.imagilan push templates/fully-auto-settings.json /sdcard/fully-auto-settings.json 
-  # adb -s tablet-$room.imagilan shell am start -a android.intent.action.REBOOT
-  # adb -s tablet-$room.imagilan shell am force-stop com.fullykiosk.emm
-  # adb -s tablet-$room.imagilan shell am start -n   com.fullykiosk.emm/de.ozerov.fully.FullyActivity
-  adb -s tablet-$room.imagilan reboot &
-  sleep 2
-  adb -s tablet-$room.imagilan disconnect
+  echo "room todo: $room"
+done
+
+for room in $rooms; do
+  if ping -c 1 -W 1 tablet-$room.imagilan &> /dev/null
+  then
+    echo "doing room: $room"
+    adb connect tablet-$room.imagilan &> /dev/null
+    adb -s tablet-$room.imagilan push templates/fully-auto-settings.json /sdcard/fully-auto-settings.json &> /dev/null
+    adb -s tablet-$room.imagilan reboot &
+    sleep 2
+    adb -s tablet-$room.imagilan disconnect &> /dev/null
+  fi
 done
